user: Typo refactor
Description:
Change the authentication folder to user folder

Expected behavior:
