+++ src/main/java/com/splitscale/fordastore/core/daos/UserRepository.java
@@ -1,6 +1,6 @@
 package com.splitscale.fordastore.core.daos;
 
-import com.splitscale.fordastore.core.authentication.User;
+import com.splitscale.fordastore.core.user.User;
 
 public interface UserRepository {
   public User findByUsername(String username);
