+++ src/main/java/com/splitscale/fordastore/core/user/SecurityService.java
@@ -1,4 +1,4 @@
-package com.splitscale.fordastore.core.authentication;
+package com.splitscale.fordastore.core.user;
 
 public interface SecurityService {
