+++ src/main/java/com/splitscale/fordastore/core/user/login/LoginInteractor.java
@@ -1,10 +1,10 @@
-package com.splitscale.fordastore.core.authentication.login;
+package com.splitscale.fordastore.core.user.login;
 
-import com.splitscale.fordastore.core.authentication.SecurityService;
-import com.splitscale.fordastore.core.authentication.User;
 import com.splitscale.fordastore.core.daos.UserRepository;
 import com.splitscale.fordastore.core.exceptions.ActionFailedException;
 import com.splitscale.fordastore.core.exceptions.ObjectNotFoundException;
+import com.splitscale.fordastore.core.user.SecurityService;
+import com.splitscale.fordastore.core.user.User;
 
 public class LoginInteractor {
   private UserRepository repository;
